requirements.txt
flask
flask-cors
numpy==1.26.4
transformers
pillow
torch
einops
timm
